# frontend/docs/bff-openapi.json
{
  "definitions": {
...
    "dto.ErrorResponse": {
      "properties": {
        "error": {
          "type": "integer"
        }
      },
      "type": "object"
    },
...
    "fleetpb.Aircraft": {
      "properties": {
        "base_airport_id": {
          "type": "string"
        },
        "current_maintenance_points": {
          "type": "number"
        },
        "current_owner_id": {
          "type": "string"
        },
        "fh_since_last_d_check": {
          "type": "number"
        },
        "id": {
          "type": "string"
        },
        "in_service": {
          "type": "boolean"
        },
        "manufactured_at": {
          "type": "string"
        },
        "max_maintenance_points_cached": {
          "type": "number"
        },
        "status": {
          "type": "string"
        },
        "tail_number": {
          "type": "string"
        },
        "total_cycles": {
          "type": "number"
        },
        "total_flight_hours": {
          "type": "number"
        },
        "type_id": {
          "type": "string"
        }
      },
      "type": "object"
    },
    "fleetpb.AircraftType": {
      "properties": {
        "base_maintenance_points": {
          "type": "number"
        },
        "base_turnaround_points": {
          "type": "number"
        },
        "characteristics": {
          "type": "string"
        },
        "cruising_speed_kph": {
          "type": "number"
        },
        "d_check_interval_fh": {
          "type": "number"
        },
        "d_check_interval_years": {
          "type": "number"
        },
        "d_check_overdue_multiplier": {
          "type": "number"
        },
        "fuel_consumption_per_hour": {
          "type": "number"
        },
        "iata_code": {
          "type": "string"
        },
        "icao_code": {
          "type": "string"
        },
        "id": {
          "type": "string"
        },
        "image_upload_id": {
          "type": "string"
        },
        "maint_cost_per_flight_hour": {
          "type": "number"
        },
        "maint_cost_per_landing": {
          "type": "number"
        },
        "maint_cost_per_takeoff": {
          "type": "number"
        },
        "manufacturer_id": {
          "type": "string"
        },
        "max_planned_seat_capacity": {
          "type": "number"
        },
        "max_range_km": {
          "type": "number"
        },
        "min_runway_length_m": {
          "type": "number"
        },
        "model_name": {
          "type": "string"
        },
        "mtow_kg": {
          "type": "number"
        },
        "price_per_unit": {
          "type": "number"
        },
        "production_points_price": {
          "type": "number"
        }
      },
      "type": "object"
    },
    "fleetpb.CreateAircraftRequest": {
      "properties": {
        "aircraft_type_id": {
          "type": "string"
        },
        "base_airport_id": {
          "type": "string"
        },
        "current_owner_id": {
          "type": "string"
        },
        "tail_number": {
          "type": "string"
        }
      },
      "type": "object"
    },
    "fleetpb.CreateAircraftResponse": {
      "properties": {
        "id": {
          "type": "string"
        }
      },
      "type": "object"
    },
    "fleetpb.CreateAircraftTypeRequest": {
      "properties": {
        "base_maintenance_points": {
          "type": "number"
        },
        "base_turnaround_points": {
          "type": "number"
        },
        "characteristics": {
          "type": "string"
        },
        "cruising_speed_kph": {
          "type": "number"
        },
        "d_check_interval_fh": {
          "type": "number"
        },
        "d_check_interval_years": {
          "type": "number"
        },
        "d_check_overdue_multiplier": {
          "type": "number"
        },
        "fuel_consumption_per_hour": {
          "type": "number"
        },
        "iata_code": {
          "type": "string"
        },
        "icao_code": {
          "type": "string"
        },
        "image_upload_id": {
          "type": "string"
        },
        "maint_cost_per_flight_hour": {
          "type": "number"
        },
        "maint_cost_per_landing": {
          "type": "number"
        },
        "maint_cost_per_takeoff": {
          "type": "number"
        },
        "manufacturer_id": {
          "type": "string"
        },
        "max_planned_seat_capacity": {
          "type": "number"
        },
        "max_range_km": {
          "type": "number"
        },
        "min_runway_length_m": {
          "type": "number"
        },
        "model_name": {
          "type": "string"
        },
        "mtow_kg": {
          "type": "number"
        },
        "price_per_unit": {
          "type": "number"
        },
        "production_points_price": {
          "type": "number"
        }
      },
      "type": "object"
    },
    "fleetpb.ListAircraftTypesResponse": {
      "properties": {
        "items": {
          "items": {
            "$ref": "#/definitions/fleetpb.AircraftType"
          },
          "type": "array"
        }
      },
      "type": "object"
    },
    "fleetpb.ListAircraftsResponse": {
      "properties": {
        "items": {
          "items": {
            "$ref": "#/definitions/fleetpb.Aircraft"
          },
          "type": "array"
        }
      },
      "type": "object"
    },
...
    "bff.FleetAircraftDetailResponse": {
      "type": "object"
    },
    "bff.FleetAircraftResponse": {
      "type": "object"
    },
...
    "bff.FleetPurchaseResponse": {
      "type": "object"
    },
...
  },
  "host": "localhost:4200",
  "info": {
    "contact": {
      "email": "lboris201@yandex.ru",
      "name": "Jeno7u",
      "url": "https://github.com/Jeno7u"
    },
    "title": "AirlineSim BFF API",
    "version": "1.0",
    "description": "Backend-for-frontend API. Based on backend OpenAPI and extended with BFF cache/filter query parameters."
  },
  "paths": {
    "/aircraft": {
      "post": {
        "consumes": [
          "application/json"
        ],
        "description": "Purchases a new aircraft from a selected aircraft type for the authenticated user",
        "parameters": [
          {
            "description": "Aircraft details",
            "in": "body",
            "name": "request",
            "required": true,
            "schema": {
              "$ref": "#/definitions/fleetpb.CreateAircraftRequest"
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "201": {
            "description": "Aircraft purchased successfully, id returned",
            "schema": {
              "$ref": "#/definitions/fleetpb.CreateAircraftResponse"
            }
          },
          "400": {
            "description": "1 - request validation error, 2 - aircraft type not found",
            "schema": {
              "$ref": "#/definitions/dto.ErrorResponse"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "summary": "Purchase aircraft",
        "tags": [
          "Aircraft"
        ]
      }
    },
    "/aircraft-types": {
      "get": {
        "consumes": [
          "application/json"
        ],
        "description": "Returns a list of available aircraft types\n\nBFF caches this list in memory, retries safe backend failures within BFF_BACKEND_MAX_SAFE_ATTEMPTS, and supports exact field filters plus q search across text fields.",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/fleetpb.ListAircraftTypesResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/dto.ErrorResponse"
            }
          }
        },
        "summary": "List aircraft types",
        "tags": [
          "Aircraft"
        ],
        "parameters": [
          {
            "description": "Search query applied to all string fields.",
            "in": "query",
            "name": "q",
            "required": false,
            "type": "string"
          },
          {
            "description": "Force BFF to refresh its in-memory cache from backend.",
            "enum": [
              "true",
              "false"
            ],
            "in": "query",
            "name": "refresh",
            "required": false,
            "type": "string"
          },
          {
            "description": "Exact BFF filter by base_maintenance_points.",
            "in": "query",
            "name": "base_maintenance_points",
            "required": false,
            "type": "number"
          },
          {
            "description": "Exact BFF filter by base_turnaround_points.",
            "in": "query",
            "name": "base_turnaround_points",
            "required": false,
            "type": "number"
          },
          {
            "description": "Exact BFF filter by characteristics.",
            "in": "query",
            "name": "characteristics",
            "required": false,
            "type": "string"
          },
          {
            "description": "Exact BFF filter by cruising_speed_kph.",
            "in": "query",
            "name": "cruising_speed_kph",
            "required": false,
            "type": "number"
          },
          {
            "description": "Exact BFF filter by d_check_interval_fh.",
            "in": "query",
            "name": "d_check_interval_fh",
            "required": false,
            "type": "number"
          },
          {
            "description": "Exact BFF filter by d_check_interval_years.",
            "in": "query",
            "name": "d_check_interval_years",
            "required": false,
            "type": "number"
          },
          {
            "description": "Exact BFF filter by d_check_overdue_multiplier.",
            "in": "query",
            "name": "d_check_overdue_multiplier",
            "required": false,
            "type": "number"
          },
          {
            "description": "Exact BFF filter by fuel_consumption_per_hour.",
            "in": "query",
            "name": "fuel_consumption_per_hour",
            "required": false,
            "type": "number"
          },
          {
            "description": "Exact BFF filter by iata_code.",
            "in": "query",
            "name": "iata_code",
            "required": false,
            "type": "string"
          },
          {
            "description": "Exact BFF filter by icao_code.",
            "in": "query",
            "name": "icao_code",
            "required": false,
            "type": "string"
          },
          {
            "description": "Exact BFF filter by id.",
            "in": "query",
            "name": "id",
            "required": false,
            "type": "string"
          },
          {
            "description": "Exact BFF filter by image_upload_id.",
            "in": "query",
            "name": "image_upload_id",
            "required": false,
            "type": "string"
          },
          {
            "description": "Exact BFF filter by maint_cost_per_flight_hour.",
            "in": "query",
            "name": "maint_cost_per_flight_hour",
            "required": false,
            "type": "number"
          },
          {
            "description": "Exact BFF filter by maint_cost_per_landing.",
            "in": "query",
            "name": "maint_cost_per_landing",
            "required": false,
            "type": "number"
          },
          {
            "description": "Exact BFF filter by maint_cost_per_takeoff.",
            "in": "query",
            "name": "maint_cost_per_takeoff",
            "required": false,
            "type": "number"
          },
          {
            "description": "Exact BFF filter by manufacturer_id.",
            "in": "query",
            "name": "manufacturer_id",
            "required": false,
            "type": "string"
          },
          {
            "description": "Exact BFF filter by max_planned_seat_capacity.",
            "in": "query",
            "name": "max_planned_seat_capacity",
            "required": false,
            "type": "number"
          },
          {
            "description": "Exact BFF filter by max_range_km.",
            "in": "query",
            "name": "max_range_km",
            "required": false,
            "type": "number"
          },
          {
            "description": "Exact BFF filter by min_runway_length_m.",
            "in": "query",
            "name": "min_runway_length_m",
            "required": false,
            "type": "number"
          },
          {
            "description": "Exact BFF filter by model_name.",
            "in": "query",
            "name": "model_name",
            "required": false,
            "type": "string"
          },
          {
            "description": "Exact BFF filter by mtow_kg.",
            "in": "query",
            "name": "mtow_kg",
            "required": false,
            "type": "number"
          },
          {
            "description": "Exact BFF filter by price_per_unit.",
            "in": "query",
            "name": "price_per_unit",
            "required": false,
            "type": "number"
          },
          {
            "description": "Exact BFF filter by production_points_price.",
            "in": "query",
            "name": "production_points_price",
            "required": false,
            "type": "number"
          }
        ]
      },
      "post": {
        "consumes": [
          "application/json"
        ],
        "description": "Create a new aircraft type (admin only)",
        "parameters": [
          {
            "description": "Aircraft type details",
            "in": "body",
            "name": "request",
            "required": true,
            "schema": {
              "$ref": "#/definitions/fleetpb.CreateAircraftTypeRequest"
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "201": {
            "description": "Created",
            "schema": {
              "$ref": "#/definitions/fleetpb.AircraftType"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/dto.ErrorResponse"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/dto.ErrorResponse"
            }
          }
        },
        "summary": "Create aircraft type",
        "tags": [
          "Aircraft"
        ]
      }
    },
    "/aircraft-types/{id}": {
      "get": {
        "consumes": [
          "application/json"
        ],
        "description": "Returns details for a single aircraft type",
        "parameters": [
          {
            "description": "Aircraft type id",
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string"
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/fleetpb.AircraftType"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/dto.ErrorResponse"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/dto.ErrorResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/dto.ErrorResponse"
            }
          }
        },
        "summary": "Get aircraft type",
        "tags": [
          "Aircraft"
        ]
      }
    },
    "/aircraft/{id}": {
      "get": {
        "consumes": [
          "application/json"
        ],
        "description": "Returns details for a single aircraft owned by the airline",
        "parameters": [
          {
            "description": "Aircraft id",
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string"
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/fleetpb.Aircraft"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/dto.ErrorResponse"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/dto.ErrorResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/dto.ErrorResponse"
            }
          }
        },
        "summary": "Get aircraft",
        "tags": [
          "Aircraft"
        ]
      },
      "patch": {
        "consumes": [
          "application/json"
        ],
        "description": "Update aircraft tail number (owner only)",
        "parameters": [
          {
            "description": "Aircraft id",
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string"
          },
          {
            "description": "{\\",
            "in": "body",
            "name": "request",
            "required": true,
            "schema": {
              "type": "object"
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/fleetpb.Aircraft"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/dto.ErrorResponse"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "409": {
            "description": "Tail number conflict"
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/dto.ErrorResponse"
            }
          }
        },
        "summary": "Update aircraft tail number",
        "tags": [
          "Aircraft"
        ]
      }
    },
    "/aircrafts": {
      "get": {
        "consumes": [
          "application/json"
        ],
        "description": "Returns list of aircrafts owned by the authenticated airline",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/fleetpb.ListAircraftsResponse"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/dto.ErrorResponse"
            }
          }
        },
        "summary": "List aircrafts for airline",
        "tags": [
          "Aircraft"
        ]
      }
    },
...
    "/fleet/aircraft": {
      "get": {
        "description": "Returns the enriched owned fleet list with product-facing empty state.",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "schema": {
              "$ref": "#/definitions/bff.FleetAircraftResponse"
            }
          },
          "400": {},
          "401": {},
          "404": {},
          "409": {},
          "500": {}
        }
      },
      "post": {
        "description": "Repeats purchase-preview checks, calls backend POST /aircraft once, clears relevant BFF caches and returns the enriched purchased aircraft plus next action.",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "schema": {
              "$ref": "#/definitions/bff.FleetPurchaseResponse"
            }
          },
          "400": {},
          "401": {},
          "404": {},
          "409": {},
          "500": {}
        }
      }
    },
    "/fleet/aircraft/{id}": {
      "get": {
        "description": "Returns an enriched owned aircraft detail card for Fleet & Ops.",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "schema": {
              "$ref": "#/definitions/bff.FleetAircraftDetailResponse"
            }
          },
          "400": {},
          "401": {},
          "404": {},
          "409": {},
          "500": {}
        }
      }
    },
...
  },
  "schemes": [
    "http"
  ],
  "swagger": "2.0"
}